refactor: Management Report 도메인 ServiceDTO 분리(#225)
